WEBSCRAPING/requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
exceptiongroup==1.1.2
fake-useragent==1.1.3
google-api-core==2.11.1
google-api-python-client==2.94.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
h11==0.14.0
httplib2==0.22.0
idna==3.4
importlib-resources==6.0.0
lxml==4.9.3
oauthlib==3.2.2
outcome==1.2.0
packaging==23.1
pdfkit==1.0.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
pyparsing==3.1.0
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
undetected-chromedriver==3.5.0
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==2.0.4
websockets==11.0.3
wsproto==1.2.0
zipp==3.16.2
